pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smfs-qa"
version = "1.0.0"
description = "Quality audit test suite for Sonic Market Feed Service"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=12.0",
    "pydantic>=2.0",
    "solders>=0.21",
    "solana>=0.34",
    "tenacity>=8.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-rerunfailures>=14.0",
    "pytest-timeout>=2.0",
    "pytest-xdist>=3.5",
    "allure-pytest>=2.13",
]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
]
load = [
    "locust>=2.20",
    "websocket-client>=1.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 30
addopts = "--strict-markers -v"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(message)s"
markers = [
    "finding: marks tests that document API findings/anomalies",
    "flaky: marks tests with non-deterministic behavior",
    "perf: marks performance tests",
    "rest: marks REST API tests",
    "websocket: marks WebSocket tests",
    "solana: marks Solana stream tests",
    "e2e: marks end-to-end integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src/smfs_qa"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["N815"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
